backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic-settings==2.2.1
boto3==1.34.48
sqlmodel==0.0.16
alembic==1.13.1
psycopg2-binary==2.9.9
openai==1.12.0
PyYAML==6.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
